fix : 좋아요 분산락 적용

+++ src/main/java/com/sparta/restplaceforj/service/LikeService.java
@@ -7,10 +7,14 @@ import com.sparta.restplaceforj.entity.CommentLike;
 import com.sparta.restplaceforj.entity.Post;
 import com.sparta.restplaceforj.entity.PostLike;
 import com.sparta.restplaceforj.entity.User;
+import com.sparta.restplaceforj.exception.CommonException;
+import com.sparta.restplaceforj.exception.ErrorEnum;
+import com.sparta.restplaceforj.provider.RedisProvider;
 import com.sparta.restplaceforj.repository.CommentLikeRepository;
 import com.sparta.restplaceforj.repository.CommentRepository;
 import com.sparta.restplaceforj.repository.PostLikeRepository;
 import com.sparta.restplaceforj.repository.PostRepository;
+import com.sparta.restplaceforj.repository.UserRepository;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -25,6 +29,7 @@ public class LikeService {
   private final PostRepository postRepository;
   private final CommentRepository commentRepository;
   private final CommentLikeRepository commentLikeRepository;
+  private final RedisProvider redisProvider;
 
   /**
    * 글 좋아요
@@ -34,28 +39,37 @@ public class LikeService {
    * @return PostLikeResponseDto 필드명 : id, userId, postId
    */
   public Optional<PostLikeResponseDto> createPostLike(long postId, User user) {
-    Post post = postRepository.findByIdOrThrow(postId);
-    PostLike postLike = PostLike.builder()
-        .user(user)
-        .post(post)
-        .build();
-
-    if (postLikeRepository.existsByPostAndUser(post, user)) {
-      post.removeLikeFromPost();
-      postRepository.save(post);
-      PostLike findPostLike = postLikeRepository.findByPostLikeOrThrow(post, user);
-      postLikeRepository.delete(findPostLike);
-      return Optional.empty();
+    String lockKey = "post_like_" + postId + "_" + user.getId();
+    boolean lock = redisProvider.validLock(lockKey, 10);
+    if (!lock) {
+      throw new CommonException(ErrorEnum.VALID_ERROR);
     }
+    try {
+      Post post = postRepository.findByIdOrThrow(postId);
+      PostLike postLike = PostLike.builder()
+          .user(user)
+          .post(post)
+          .build();
 
-    post.addLikeToPost();
-    postRepository.save(post);
-    PostLike savedPostLike = postLikeRepository.save(postLike);
-    PostLikeResponseDto postLikeResponseDto = PostLikeResponseDto.builder()
-        .postLike(savedPostLike)
-        .build();
+      if (postLikeRepository.existsByPostAndUser(post, user)) {
+        post.removeLikeFromPost();
+        postRepository.save(post);
+        PostLike findPostLike = postLikeRepository.findByPostLikeOrThrow(post, user);
+        postLikeRepository.delete(findPostLike);
+        return Optional.empty();
+      }
 
-    return Optional.of(postLikeResponseDto);
+      post.addLikeToPost();
+      postRepository.save(post);
+      PostLike savedPostLike = postLikeRepository.save(postLike);
+      PostLikeResponseDto postLikeResponseDto = PostLikeResponseDto.builder()
+          .postLike(savedPostLike)
+          .build();
+
+      return Optional.of(postLikeResponseDto);
+    } finally {
+      redisProvider.returnLock(lockKey);
+    }
   }
 
   /**
@@ -66,30 +80,40 @@ public class LikeService {
    * @return CommentLikeResponseDto 필드명 : id, userId, commentId
    */
   public Optional<CommentLikeResponseDto> createCommentLike(long commentId, User user) {
-    Comment comment = commentRepository.findByIdOrThrow(commentId);
-    CommentLike commentLike = CommentLike.builder()
-        .user(user)
-        .comment(comment)
-        .build();
+    String lockKey = "commentLike:" + commentId + ":" + user.getId();
+    boolean lock = redisProvider.validLock(lockKey, 10);
 
-    if (commentLikeRepository.existsByCommentAndUser(comment, user)) {
-      comment.removeLikeFromComment();
+    if (!lock) {
+      throw new CommonException(ErrorEnum.VALID_ERROR);
+    }
+    try {
+      Comment comment = commentRepository.findByIdOrThrow(commentId);
+      CommentLike commentLike = CommentLike.builder()
+          .user(user)
+          .comment(comment)
+          .build();
+
+      if (commentLikeRepository.existsByCommentAndUser(comment, user)) {
+        comment.removeLikeFromComment();
+        commentRepository.save(comment);
+
+        CommentLike findCommentLike = commentLikeRepository.findByCommentLikeOrThrow(comment,
+            user);
+        commentLikeRepository.delete(findCommentLike);
+        return Optional.empty();
+      }
+
+      comment.addLikeToComment();
       commentRepository.save(comment);
 
-      CommentLike findCommentLike = commentLikeRepository.findByCommentLikeOrThrow(comment,
-          user);
-      commentLikeRepository.delete(findCommentLike);
-      return Optional.empty();
+      CommentLike savedCommentLike = commentLikeRepository.save(commentLike);
+      CommentLikeResponseDto commentLikeResponseDto = CommentLikeResponseDto.builder()
+          .commentLike(savedCommentLike)
+          .build();
+
+      return Optional.of(commentLikeResponseDto);
+    } finally {
+      redisProvider.returnLock(lockKey);
     }
-
-    comment.addLikeToComment();
-    commentRepository.save(comment);
-
-    CommentLike savedCommentLike = commentLikeRepository.save(commentLike);
-    CommentLikeResponseDto commentLikeResponseDto = CommentLikeResponseDto.builder()
-        .commentLike(savedCommentLike)
-        .build();
-
-    return Optional.of(commentLikeResponseDto);
   }
 }
